% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi_createAOI.r
\name{aoi_createAOI}
\alias{aoi_createAOI}
\title{Create Area of Interest}
\usage{
aoi_createAOI(
  x,
  method = c("ch", "env", "raster", "river"),
  buffer = 0,
  outF = c("shp", "kml", "tif", "rdata"),
  res = 1000,
  mask = FALSE,
  crs,
  dsn
)
}
\arguments{
\item{x}{Character vector or sf object. If character, then path(s) (optionally relative to working directory) to shapefile or kml of spatial points or polygon(s). If multiple files are used, then the aoi is the union of all these files.}

\item{method}{Character. Method to create aoi. One of, ch (convex hull around points), env (rectangular envelope), r (raster template)
or river (tbc). Note \code{ch} and \code{env} may be extended by a \code{buffer} around them.}

\item{buffer}{Numeric scalar. Buffer distance (in metres) around points or polygon to create aoi, default is no buffer.}

\item{outF}{Character vector indicating output format for aoi. Can be one ore more of shp (shapefile), kml, rdata (sf object saved as .rdata)}

\item{res}{Numeric. Expected resolution of raster covariates (useful to have \code{env} and raster template aligned)}

\item{mask}{Logical. If TRUE, will create an additional raster masked outside the aoi}

\item{crs}{Force a particular projection on spatial input \code{x}, otherwise projected to corresponding UTM zone}

\item{dsn}{Character vector with filepath to folder to store output aoi. Either length one or equal to outF with a path for each format.}
}
\value{
Returns an sf polygon object of the area of interest or a spatRaster object, depending on \code{method} requested (and optionally saves this to disk in requested format)
}
\description{
\code{aoi_createAOI} creates an area of interest from sample points or existing polygons and optionally, projects this to an appropriate UTM zone.
}
