% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi_createMask.r
\name{aoi_createMask}
\alias{aoi_createMask}
\title{Create a mask for areas that are unavailable for sampling}
\usage{
aoi_createMask(vData, rData, buffer, maskValue = 1, background = NA)
}
\arguments{
\item{vData}{Either a single sf object or a list of point, line or polygon sf objects that will optionally be buffered, and converted to a raster masked area.}

\item{rData}{Terra \code{SpatRaster} with required resolution, extent and projection with one or more raster layers. Any value above 0 or NA will be used as a mask. If no rasters are required for the mask, then at least one layer must be provided as a template, typically the result of \code{aoi_createAOI(method = "r")} function. If used as a template, the layers does not need to contain any values, but any NAs in this template will be added to the mask.}

\item{buffer}{Either a single number, representing a universal buffer (in projection units, e.g. metres) to be applied to all vector data in \code{vData} or a vector of numeric buffer values corresponding to the order of the data in \code{vData}.}

\item{maskValue}{Numeric. Value to use for masked area. Defaults to 1.}

\item{background}{Numeric. Value to use as background (unmasked area). Defaults to NA.}
}
\value{
A terra \code{SpatRaster} with two values (or single value and NAs) representing area to mask.
}
\description{
\code{aoi_createMask()} creates a mask using vector and raster layers (e.g. roads, streams, settlements, specific habitat types or custom polygons) that will be discounted from sampling site selection.
}
\details{
Any areas included in the (buffered) point, line or polygon files, as well as the supplied rasters will be combined into a single binary raster mask, representing the union of all layers. It will have the spatial extent, resolution and projecton of the layers in \code{rData}. All sf objects and rasters should be in the same projection, otherwise they will be reprojected to the projection of the raster data provided. Masks used in \code{sd_create_design} show areas to exclude (with 0=invalid; 1=valid).
}
\examples{
\dontrun{
# How to specify buffers for each sf object:
vData <- list(points, lines, polygons)
buffer <- c(1000, 30, 0)
}
}
