% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eo_AppEEARS_functions.r
\name{eo_authAppEEARS}
\alias{eo_authAppEEARS}
\title{API AppEEARS - Authentication}
\usage{
eo_authAppEEARS()
}
\value{
API token is returned (valid for 48 hours)
}
\description{
Authenticate API
Enter NASA EARTH EXPLORER username and password at prompt to access API
}
