% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eo_autoDEM.r
\name{eo_autoDEM}
\alias{eo_autoDEM}
\title{Automated one step download of DEMs from LPDAAC}
\usage{
eo_autoDEM(
  sf,
  products = c("ASTER", "SRTM"),
  taskName,
  token,
  outDir,
  outDir2,
  projTo,
  hillshade = TRUE,
  tifOnly = TRUE,
  write = TRUE,
  out = c("proj", "fp", "wgs"),
  ...
)
}
\arguments{
\item{sf}{filepath to spatial file (readable by sf package, eg .shp, or .kml) or sf object in R}

\item{products}{Character vector. Product names, can be shortened to 'ASTER' and/or 'SRTM', or full ProductandVersion names (see products in \code{eo_getAppEEARS()})}

\item{taskName}{Character vector. Name to identify task.}

\item{token}{Authentication token. Output of \code{authAppEEARS()}}

\item{outDir}{Path to directory where data is downloaded}

\item{outDir2}{Alternative output folder for projected and hillshade rasters. Defaults to \code{outDir}}

\item{projTo}{SpatRaster template for projection, as output by \code{aoi_createRaster()} - containing, desired, extent, resolution and projection information; or coordinate reference information. See \code{?terra::project()} for more details. Coordinate reference formats: WKT, PROJ.4 (e.g., +proj=longlat +datum=WGS84), or an EPSG code (e.g., "epsg:4326").}

\item{hillshade}{Logical. If \code{TRUE}, hillshade is also created from projected raster. Further arguments to \code{terra::shade()} may be included in ...}

\item{tifOnly}{Logical. If \code{TRUE}, only principal tif files will be downloaded for each product.}

\item{write}{Logical. Should processed rasters be written to disk.}

\item{out}{Character vector. Output format. Either 'fp' to return file paths of downloaded rasters; 'wgs' to return only rasters in native projection; 'proj' to return projected (and hillshade) rasters (only if projTo is included)}

\item{...}{Further argumens to \code{terra::shade()}}
}
\value{
Either a character vector of filenames written to disk, or a SpatRaster object containing either the downloaded rasters in native projection, or the processed rasters
}
\description{
Download a digital elevation product (Aster or SRTM) from LPDAAC cropped to given extent, and optionally, transform to a utm projection, and create a hillshade raster. Other products (see \code{eo_getProducts()}) are also possible to download and project with this function.
}
\details{
\emph{Note} The function will wait until the LPDAAC task has completed. If this is taking a long time, cancel and run \code{eo_getStatus()} without arguments to see the status of the most recent task. When this shows as 'done', the files can be downloaded with \code{eo_getData()}.
}
\seealso{
\code{\link{eo_getProduct}} to search all downloadable layers, \code{\link{eo_getAppEEARS }} to request a data download from the AppEAEARS API, \code{\link{eo_getStatus}} to query the status of a data request.
}
