% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eo_AppEEARS_functions.r
\name{eo_deleteTask}
\alias{eo_deleteTask}
\title{Delete a task}
\usage{
eo_deleteTask(task_id, token)
}
\arguments{
\item{token}{Authentication token. Output of \code{authAppEEARS()}}
}
\description{
Delete a task
}
