% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eo_AppEEARS_functions.r
\name{eo_getAppEEARS}
\alias{eo_getAppEEARS}
\title{LP DAAC Data Download
Download}
\usage{
eo_getAppEEARS(
  sf,
  products,
  proj = "native",
  taskName,
  outFormat = c("geotiff", "netcdf4"),
  startDate,
  endDate,
  recurring,
  fromYear,
  toYear,
  token,
  verbose = TRUE
)
}
\arguments{
\item{sf}{filepath to spatial file (readable by sf package, eg .shp, or .kml) or sf object in R}

\item{proj}{Character vector. Projection for requested data (see \code{getProjections()} for options), e.g. "native", or "geographic", "easegrid_2_global"}

\item{taskName}{Character vector. Name to identify task.}

\item{outFormat}{Character vector. File format for output. Either geotiff or netcdf4}

\item{startDate}{Start date for data series. If \code{recurring} is FALSE in format \%m-\%d-\%Y e.g. '12-31-2020'; if recurring is TRUE, then in format \%m-\%d e.g. '12-31'}

\item{endDate}{End date for data series. See above for format.}

\item{recurring}{Logical. Are data to be downloaded for recurring time within start and end date. e.g. at monthly frequency. See below.}

\item{fromYear}{Only if recurring is TRUE. Start year for recurring series.}

\item{toYear}{Only if recurring is TRUE. End year for recurring series.}

\item{token}{Authentication token. Output of \code{authAppEEARS()}}

\item{product}{Data frame of products and layers (see \code{getProducts()} for options)}
}
\value{
json with task ID and status (can be used to query status later)
}
\description{
LP DAAC Data Download
Download
}
