% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eo_AppEEARS_functions.r
\name{eo_getData}
\alias{eo_getData}
\title{Data download}
\usage{
eo_getData(outDir, task_id, download = FALSE, name, limit = "ASTGTM_NC")
}
\arguments{
\item{outDir}{Path to directory where data is downloaded}

\item{name}{Optional character string to append to downloaded files}

\item{limit}{Character vector of (partial) product names to limit downloaded files to}
}
\value{
A vector of filepaths to the downloaded files or nothing if download is \code{FALSE}
}
\description{
Data download
}
