% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eo_AppEEARS_functions.r
\name{eo_getProduct}
\alias{eo_getProduct}
\title{Get list of products and layers}
\usage{
eo_getProduct(products = NULL)
}
\arguments{
\item{products}{Character vector with 'ProductAndVersion' of required layer. e.g. for Aster DEM v003: "ASTGTM_NC.003" Names taken from product list obtained from \code{getProduct()} called without arguments.}
}
\value{
If products is NULL, a dataframe of all available products and associated information, if specific products are requested, then a dataframe of product and layer names
}
\description{
Get dataframe of products available from API. see also https://lpdaac.usgs.gov/product_search/. Call without arguments to get the product dataframe. Then use product IDs from the 'ProductAndVersion' column to get specific layers. The resulting data frame can be subsetted (if necessary) and used as input to \code{getAppEEARS()}
}
\examples{
search for specific products in dataframe produced from this function.
prod.df <- getProduct()
prod.df[grepl("dem", prod.df$Description, ignore.case = T),]
prod.df[grepl("AST", prod.df$Product, ignore.case = T),]
}
