% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eo_AppEEARS_functions.r
\name{eo_getProjection}
\alias{eo_getProjection}
\title{Get projection info}
\usage{
eo_getProjection(silent = FALSE)
}
\arguments{
\item{silent}{Logical. Print projection information}
}
\value{
A data frame of names, descriptions and corresponding platforms for where projections are available (invisibly). Optionally prints the data frame to the console.
}
\description{
Get available projection information corresponding to products and platforms
}
