% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eo_AppEEARS_functions.r
\name{eo_getStatus}
\alias{eo_getStatus}
\title{Monitor the status of the task}
\usage{
eo_getStatus(time = NULL, n = 1, task_id, token)
}
\arguments{
\item{time}{Time (in seconds) for repeated checking of status. For single call, use NULL. Currently only checks repeatedly for most recent task.}

\item{n}{How many recent tasks to check status of?}

\item{task_id.}{Optional. Specify a particular task to check status of}
}
\description{
Get the status of a previously submitted task
}
