% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_getData.r
\name{osm_getData}
\alias{osm_getData}
\title{Extract layers from Open Street Maps}
\usage{
osm_getData(
  x,
  key = c("highway", "place", "landuse", "waterway"),
  tags,
  cols,
  dsn = NULL,
  suffix = NULL,
  ...
)
}
\arguments{
\item{x}{sf object from which an extent can be extracted. OSM data is downloaded within this bouding box.}

\item{key}{Character. OSM feature to download. e.g. highway for roads, place for settlements, waterway for rivers. See (https://wiki.openstreetmap.org/wiki/Map_Features) for details.}

\item{tags}{Character vector. Tags to subset an OSM feature. If missing, then default for key = "place" are c("city","town","village","hamlet","isolated_dwelling"), NULL for key = "highway", "river" for key = "waterway".}

\item{cols}{Character vector. Column names to subset result. Only these columns will be returned in osm extract. Must match column names in osm data. e.g. Recommended columns to keep:
highways: c("osm_id", "name", "highway", "surface")
places: c("osm_id", "name", "place")}

\item{dsn}{Character. Directory path to save osm data as shapefile. Will be saved in same projection as \code{x}. If NULL, then no file is saved.}

\item{suffix}{Character. Suffix to append to filename of shapefile.}

\item{...}{Further arguments to \code{sf::st_write()}}
}
\value{
If a path is supplied (\code{dsn}), then osm data will be saved. Invisibly returns an sf object of the osm data (e.g. points or lines).
}
\description{
Extract layers such as roads and settlements from OSM. OSM layers will be downloaded and cropped to the extent supplied in the function. OSM data can be filtered by specifying tags within osm features (e.g. cities within places) and/or by removing columns in the resulting data.
}
