% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osm_getDist.r
\name{osm_getDist}
\alias{osm_getDist}
\title{Create distance to Feature raster}
\usage{
osm_getDist(x, y, key, cols, outDir = NULL, outDir2 = NULL, rasterize = TRUE)
}
\arguments{
\item{x}{SpatRaster. From aoi_getAOI() - provides extent, resolution and
projection for resulting distance raster.}

\item{y}{sf object. Alternative source of feature (e.g. roads), will be used
instead of osm data. If present, key, cols are ignored.}

\item{key}{Character. OSM feature to download. e.g. highway for roads, place for settlements, waterway for rivers. See (https://wiki.openstreetmap.org/wiki/Map_Features) for details.}

\item{cols}{Character vector. Column names to subset result. Only these columns will be returned in osm extract. Must match column names in osm data. e.g. Recommended columns to keep:
highways: c("osm_id", "name", "highway", "surface")
places: c("osm_id", "name", "place")}

\item{outDir}{Character. Path to folder to save osm vector data (optional).}

\item{outDir2}{Character. Path to folder to save resulting distance raster
(optional).}

\item{rasterize}{Logical. Rasterize before distance calculation. Increases
speed of operation - see ?terra::distance()}
}
\value{
SpatRaster with distance to feature (in m), invisibly.
}
\description{
Creates distance rasters from osm feature data, e.g. roads, rivers or a
supplied feature data set (sf object).
}
